Revert "[WebConsole] Add closed loop polling to performance tab graphs"
This reverts commit 48c5b7764900084d4e9b20fcc4dd143de2f43fb5.
